refactor: dynamic route for posts
