+++ src/app/(posts)/[id]/page.tsx
@@ -1,25 +1,25 @@
-import { readPostIds, readPostData } from "../../../utils/service";
 import { MDXRemote } from "next-mdx-remote/rsc";
-// import ProfileCard from "@/app/components/ProfileCard";
 import Article from "@/app/components/Article";
-import { Suspense, use } from "react";
-import { notFound } from "next/navigation";
+import { Suspense } from "react";
 import dynamic from "next/dynamic";
-import { getPostIDs } from "@/utils/orch";
+import { getPostIDs, getPostData } from "@/utils/orch";
 
 const MarkdownImage = dynamic(() => import("@/app/components/MarkdownImage"));
 
 const components = { MarkdownImage };
 
-const Page = ({ params }: { params: { id: string } }) => {
-  const validPostIds = use(getPostIDs());
+export const dynamicParams = false;
 
-  if (!validPostIds.includes(params.id)) {
-    notFound();
-  }
+export async function generateStaticParams() {
+  const postIDs = (await getPostIDs()) as string[];
 
-  const { data, content } = readPostData(params.id);
+  return postIDs.map((id) => {
+    id;
+  });
+}
 
+const Page = async ({ params }: { params: { id: string } }) => {
+  const { data, content } = await getPostData(params.id);
   return (
     <Article data={data} content={content}>
       <Suspense fallback={<>Loading..</>}>
